email_automation_app/requirements.txt
langchain-core
langchain_community 
langgraph 
langchain_google_genai
langchain_chroma
chromadb
mail-parser-reply
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
python-dotenv
sse_starlette
uvicorn
gunicorn
fastapi
